git-svn-id: https://svn.kuali.org/repos/student/branches/ks-poc-1.1.x-br@13983 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-core/ks-common-ui/src/main/java/org/kuali/student/common/ui/client/configurable/mvc/binding/HasValueBinding.java
@@ -20,7 +20,6 @@ import java.util.Date;
 import org.kuali.student.common.ui.client.mvc.DataModel;
 import org.kuali.student.core.assembly.data.Data;
 import org.kuali.student.core.assembly.data.QueryPath;
-import org.kuali.student.core.assembly.data.Data.DataValue;
 
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.user.client.ui.HasValue;
@@ -52,7 +51,7 @@ public class HasValueBinding extends ModelWidgetBindingSupport<HasValue> {
         } else if (value instanceof Double) {
             model.set(qPath, (Double) value);
         } else if (value instanceof Byte) {
-            model.set(qPath, (Integer) value);
+            model.set(qPath, ((Byte) value).intValue());
         } else if (value instanceof Boolean) {
             model.set(qPath, (Boolean) value);
         } else if (value instanceof Date) {
